backend/pyproject.toml
[project]
name = "pensieve-backend"
version = "1.0.0"
description = "Privacy-first reflective journaling system"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    
    # Database - SQLite for local dev
    "aiosqlite>=0.19.0",
    "sqlalchemy>=2.0.25",
    
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[argon2]>=1.7.4",
    "argon2-cffi>=23.1.0",
    "cryptography>=42.0.0",
    
    # Utilities
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
ml = [
    # ML & NLP (optional for full features)
    "torch>=2.1.0",
    "transformers>=4.37.0",
    "sentence-transformers>=2.3.0",
    "scikit-learn>=1.4.0",
    "spacy>=3.7.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
